+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+
+5.7.1 (2014/10/11)
+----------------
+
+WebimRoom add 'owner' property
+
+WebimException move to webim.client package
+
+
 5.7 (2014/10/10)
 ----------------